Fix: aií기록 생성 후 피드백 보완
