+++ src/main/java/com/umc/finly/domain/record/service/RecordFeedbackServiceImpl.java
@@ -1,38 +1,26 @@
 package com.umc.finly.domain.record.service;
 
-import com.umc.finly.domain.market.stock.entity.Stock;
-import com.umc.finly.domain.market.stock.repository.StockRepository;
 import com.umc.finly.domain.record.dto.RecordFeedbackRes;
 import com.umc.finly.domain.record.entity.RecordEntry;
 import com.umc.finly.domain.record.entity.RecordFeedback;
 import com.umc.finly.domain.record.enums.FeedbackStatus;
-import com.umc.finly.domain.record.infra.FeedbackPromptBuilder;
-import com.umc.finly.domain.record.infra.OpenAiFeedbackClient;
 import com.umc.finly.domain.record.repository.RecordEntryRepository;
 import com.umc.finly.domain.record.repository.RecordFeedbackRepository;
 import com.umc.finly.global.apiPayload.exception.CustomException;
 import com.umc.finly.global.apiPayload.response.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class RecordFeedbackServiceImpl implements RecordFeedbackService {
 
-    private static final int PAST_RECORDS_LIMIT = 20;
-
     private final RecordFeedbackRepository feedbackRepository;
     private final RecordEntryRepository recordEntryRepository;
-    private final StockRepository stockRepository;
-    private final OpenAiFeedbackClient openAiFeedbackClient;
-    private final FeedbackPromptBuilder promptBuilder;
+    private final RecordFeedbackAsyncExecutor asyncExecutor;
 
     @Override
     @Transactional
@@ -45,53 +33,10 @@ public class RecordFeedbackServiceImpl implements RecordFeedbackService {
 
         RecordFeedback savedFeedback = feedbackRepository.save(feedback);
 
-        generateFeedbackAsync(savedFeedback.getId(), memberId, recordEntry.getId());
+        // 별도 Bean에서 호출하여 @Async가 정상 작동하도록 함
+        asyncExecutor.generateFeedbackAsync(savedFeedback.getId(), memberId, recordEntry.getId());
 
         return savedFeedback;
-    }
-
-    @Async
-    @Transactional
-    public void generateFeedbackAsync(Long feedbackId, Long memberId, Long recordEntryId) {
-        RecordFeedback feedback = feedbackRepository.findById(feedbackId)
-                .orElse(null);
-
-        if (feedback == null) {
-            log.error("Feedback not found for id: {}", feedbackId);
-            return;
-        }
-
-        try {
-            feedback.markGenerating();
-            feedbackRepository.save(feedback);
-
-            RecordEntry currentEntry = recordEntryRepository.findById(recordEntryId)
-                    .orElseThrow(() -> new CustomException(ErrorCode.RECORD_NOT_FOUND));
-
-            Stock stock = stockRepository.findById(currentEntry.getStockId())
-                    .orElse(null);
-
-            List<RecordEntry> pastEntries = recordEntryRepository
-                    .findByMemberIdOrderByRecordDateDesc(memberId, PageRequest.of(0, PAST_RECORDS_LIMIT));
-
-            pastEntries = pastEntries.stream()
-                    .filter(e -> !e.getId().equals(recordEntryId))
-                    .toList();
-
-            String systemPrompt = promptBuilder.getSystemPrompt();
-            String userPrompt = promptBuilder.buildUserPrompt(currentEntry, stock, pastEntries);
-
-            OpenAiFeedbackClient.FeedbackResponse response = openAiFeedbackClient.generateFeedback(systemPrompt, userPrompt);
-
-            feedback.markCompleted(response.content(), response.promptTokens(), response.completionTokens());
-            feedbackRepository.save(feedback);
-
-            log.info("Feedback generated successfully for recordEntryId: {}", recordEntryId);
-        } catch (Exception e) {
-            log.error("Failed to generate feedback for recordEntryId: {}", recordEntryId, e);
-            feedback.markFailed(e.getMessage());
-            feedbackRepository.save(feedback);
-        }
     }
 
     @Override
@@ -135,7 +80,7 @@ public class RecordFeedbackServiceImpl implements RecordFeedbackService {
             feedback = feedbackRepository.save(feedback);
         }
 
-        generateFeedbackAsync(feedback.getId(), memberId, recordEntryId);
+        asyncExecutor.generateFeedbackAsync(feedback.getId(), memberId, recordEntryId);
 
         return RecordFeedbackRes.from(feedback);
     }
